% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_resume.R
\name{current_resume}
\alias{current_resume}
\title{Get current Torvik team resume table}
\usage{
current_resume()
}
\value{
A tibble containing team metrics.
}
\description{
Returns the most current Bart Torvik team results table as a tibble.
}
