% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espnGameIds.R
\name{espnGameIds}
\alias{espnGameIds}
\title{Scrape ESPN game IDs by season and date}
\usage{
espnGameIds(
  start_mmdd = "11-01",
  end_mmdd = "03-08",
  sleep_sec = 1.5,
  group = "50"
)
}
\arguments{
\item{start_mmdd}{Start date (MM-DD).}

\item{end_mmdd}{End date (MM-DD).}

\item{sleep_sec}{Base sleep time between requests.}

\item{group}{ESPN scoreboard group id (default "50").}
}
\value{
A tibble of ESPN game IDs.
}
\description{
Internal utility used to generate package data.
}
\keyword{internal}
