% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerBoxScore.R
\name{PlayerBoxScore}
\alias{PlayerBoxScore}
\title{Get player box score game log}
\usage{
PlayerBoxScore(name, season, team)
}
\arguments{
\item{name}{Player name (e.g., "Zach Edey")}

\item{season}{Season year (numeric, e.g., 2024)}

\item{team}{Team name to match}
}
\value{
A data.frame of game logs
}
\description{
Retrieves a player's season game log from Sports Reference.
}
