% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGameIds.R
\name{scrape_ncaa_game_ids}
\alias{scrape_ncaa_game_ids}
\title{Scrape NCAA game IDs by season and date}
\usage{
scrape_ncaa_game_ids(
  season_ids,
  start_mmdd = "11-01",
  end_mmdd = "03-08",
  sleep_sec = 1
)
}
\arguments{
\item{season_ids}{Vector of NCAA season division IDs}

\item{start_mmdd}{Start date (MM-DD)}

\item{end_mmdd}{End date (MM-DD)}

\item{sleep_sec}{Base sleep time between requests}
}
\value{
A tibble of game IDs
}
\description{
Internal utility used to generate package data.
}
\keyword{internal}
