% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeMachine_ratings.R
\name{timeMachine_ratings}
\alias{timeMachine_ratings}
\title{Get historical Torvik Time Machine ratings}
\usage{
timeMachine_ratings(date)
}
\arguments{
\item{date}{Date in YYYYMMDD format (must be >= 20141113)}
}
\value{
A data.frame of team ratings
}
\description{
Retrieves archived team ratings for a specific date.
}
