% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_super_sked.R
\name{get_super_sked}
\alias{get_super_sked}
\title{Get Bart Torvik super schedule}
\usage{
get_super_sked(year = as.integer(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{year}{Season year (>= 2008)}
}
\value{
A data.frame containing super schedule data
}
\description{
Retrieves the full Torvik super schedule table for a given season.
}
