% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic_team_results.R
\name{historic_team_results}
\alias{historic_team_results}
\title{Get historical Torvik team results}
\usage{
historic_team_results(year = as.integer(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{year}{Season year (must be >= 2008)}
}
\value{
A data.frame of team results
}
\description{
Retrieves the full team results table for a given season year.
}
